1.当环境的ftl文件不存在时，默认使用dev.ftl 2.日志依赖

+++ src/main/resources/templates/dev.ftl
@@ -5,17 +5,24 @@ node('Slave_95') {
       // ${item.name}
       if (fileExists('${item.fullName}')) {
         dir('./${item.fullName}') {
+          sh "git checkout master"
           sh "git reset --hard master"
           sh 'git pull'
+          sh "git checkout ${item.version}"
         }
       } else {
         sh "git clone http://'${r'${USERNAME}'}':'${r'${PASSWORD}'}'@${item.gitUrl}"
+        dir('./${item.fullName}') {
+          sh "git checkout ${item.version}"
+        }
       }
     }
   }
   stage('mvn test and build') {
     dir('./${item.fullName}') {
-      sh 'rm -rf Jenkins'
+      if (fileExists('Jenkins')) {
+        sh 'rm -rf Jenkins'
+      }
       sh '/usr/bin/sh gradlew clean build'
     }
   }
